frontend: rename FighterUIItem component
