+++ frontend/src/lib/battle/BattleFighterCard.svelte
